Task 4/requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
PyPDF2
python-docx
faiss-cpu
